Moving userAccount factory.
